IBPM.Grid — Type

A collection of types of grids used to discretize and solve the linear system. Types of flow grids include:

MultiGrid

TODO: Quadtree

IBPM.MultiGrid — Type

MultiGrid <: Grid

A type of Grid which uses a hierarchy of spacial discretizations to recursively solve until the solution converges. While faster than brute force solving large linear systems, MultiGrid typically uses relatively high memory allocation.

IBPM.MultiGrid — Method

MultiGrid(
    h::Float64,
    boundary::NTuple{2, Tuple{Float64, Float64}};
    mg::Int = 1
)

Constructor for struct MultiGrid. Returns an instance of struct MultiGrid, see Main.IBPM.MultiGrid.

Arugments

h::Float64: Grid spacing.

boundary::NTuple{2, Tuple{Float64, Float64}}: Tuple defining the boundaries of the domain.

mg::Int: Number of domains in MultiGrid. Default: 1.