IBPM.apply_bc! — Method

!***************************************************************! !* given vorticity at edges of domain, rbc, (from larger, ! ! coarser mesh), add values to correct laplacian of ! ! vorticity on the (smaller, finer) domain, r. *! !***************************************************************!

r is a vorticity-like array of size (nx-1)×(ny-1)

IBPM.coarsify! — Method

!***************************************************************! !* given vorticity on a smaller, fine mesh, (rhs) interp. ! ! values to the center region of a larger, coarser mesh ! ! (crhs). The values outside the center region are ! ! not unmodified. Result is placed in arhs *! !***************************************************************!

IBPM.get_bc! — Method

!***************************************************************! !* given vorticity on a larger, coarser mesh, interpolate ! ! it's values to the edge of a smaller, finer mesh *! !***************************************************************!