IBPM.avg_flux — Method

Compute average fluxes across cells in linearized problem (no freestream)

Return views into working memory prob.model.work.q2

IBPM.avg_flux — Method

Compute average fluxes across cells

Return views into working memory prob.model.work.q2

IBPM.direct_product! — Method

direct_product!(fq, Q, Γ, prob::LinearizedIBProblem, lev)

Compute the base flow advection terms for the linear problem.

Needs FOUR flux-shaped arrays from working memory (one extra compared to nonlinear solver)

IBPM.direct_product! — Method

direct_product!(fq, Q, Γ, prob::IBProblem, lev)

Gather the product used in computing advection term

Note that the grid level lev isn't used here, but is required for the linearized version

IBPM.direct_product_loops! — Method

directproductloops!(fq, Q, Γ, Γbc, prob)

Helper function to compute the product of Q and Γ so that the advective term is ∇⋅fq

IBPM.nl_avg! — Method

fq = scale(Q1Γ1 + Q2*Γ2)

IBPM.nonlinear! — Method

nonlinear!(nonlin, state, Γbc, lev, prob::IBProblem)

IBPM.nonlinear_OLD! — Method

nonlinear!(nonlin, state, Γbc, lev, prob)