IBPM.setup_reg — Method

Return the interpolation matrix E, which interpolates flow quantities to the immersed boundary. This depends on the position of the body points, but not their velocities

The transpose of E is the regularization matrix which smears quantities from the IB to the flow domain.

IBPM.δh — Method

δh( rf, rb , dr )

Discrete delta function used to relate flow to structure quantities