IBPM.AdamsBashforth — Type

AdamsBashforth <: ExplicitScheme

The group of multistep methods known as Adams-Bashforth methods. Currently supports:

Two step Adams-Bashforth Method

IBPM.ExplicitScheme — Type

A collection of different explicit time stepping schemes. Types of explicit schemes include:

AdamsBashforth

TODO: RungeKuttaChebyshev

Is passed as an arugment into IBProblem, see Main.IBPM.IBProblem.

IBPM.IBMatrices — Type

Matrices that can be precomputed

Correspondence with Taira & Colonius (2007) Note that not all matrices defined here are explicitly constructed C - Basic curl operator for single-grid Call curl! function for multigrid to take into account boundary conditions C' - Transforms velocity flux to circulation: Γ = C'q E - Maps fluxes to body motion, i.e. u_B = Eq Note that H = -E' is the regularization operator A - Implicit time-stepping operator for velocity flux A = I - (dt/2/h^2)*Lap

NOTE: Most of the functionality for this is in fluid-operators/lin.jl

Constructor

IBMatrices(grid::Grid, bodies::Vector{<:Body})

Arguments

grid::Grid: Grid struct of type Grid which defines and discretizes the domain.

bodies::Array{Body, 1}: 1D array of bodies created for simulation.

IBPM.IBModel — Type

IBModel{T <: Grid, V <: Body} <: SolnModel

Contains information about simulation parameters and stores static matrices. Is created using a type of grid, simulation bodies, Reynolds number, and freestream velocity. Is passed as an arugment into IBProblem, see Main.IBPM.IBProblem.

Constructor

IBModel(
    grid::Grid,
    bodies::Array{Body, 1},
    Re::Number;
    freestream::Function,
    xc=0.0,
    yc=0.0
)

Arguments

grid::Grid: Grid struct of type Grid which defines and discretizes the domain.

bodies::Array{Body, 1}: Array of bodies.

Re::Number: Reynolds number.

freestream::Function: Free-stream velocity.

xc=0.0: x coordinate offset for calculating rotational fluxes.

yc=0.0: y coordinate offset for calculating rotational fluxes.

IBPM.SolnModel — Type

SolnModel contains information about simulation parameters and stores all static (non-time varying) matrices.

IBPM.WorkingMemory — Type

Pre-allocate memory to certain vectors that can be re-used throughout the computation process.

Constructor

WorkingMemory(grid::Grid)

Arguments

grid::Grid: Grid struct of type Grid which defines and discretizes the domain.

IBPM.AB2 — Method

AB2(dt::Float64)

Special case constructor for type::AdamsBashforth for second-order Adams-Bashforth scheme.

Arguments

dt::Float64: Time stepping interval.