IBPM.AbstractIBProblem — Type

IBProblem has the info of IBModel as well as the problem structure (e.g., the explicit time stepping scheme and information about the implicit treatment via the A and B matrices and their inverses).

Maybe some opportunity for restructuring...

Looking towards possible compatibility with DifferentialEquations.jl, this would be similar to the ODEProblem

IBPM.IBProblem — Type

IBProblem <: AbstractIBProblem

Initialize the problem structure (matrices used, bodies and simulation parameters, time steppping scheme, ...)

Note: the scheme actually speaks to the terms that are explicitly treated. This is a projection method the directly enforces the no-slip condition, so some terms are implicitly treated. This information is not contained in scheme, but in the A, Ainv, B, and Binv matrices.

IBProblem(
    grid::Grid,
    bodies::Vector{<:Body{M}} where M,
    freestream = t -> (0.0, 0.0);
    Re::Float64,
    dt::Float64,
)

Arguments

grid: Discretization of the fluid domain.

bodies: The bodies to simulate.

freestream: The freestream velocity as a function of time. freestream(t) returns a tuple of each velocity component at t. Defaults to zero.

Re: Reynolds number.

dt: The time stepsize.

IBPM.LinearizedIBProblem — Type

LinearizedIBProblem <: AbstractIBProblem

Create a linearized problem from the base_state and associated IBProblem

NOTE: The freestream value in the nonlinear IBProblem will be incorrect for the linearized case, but this field is not used in base_flux!(..., prob::LinearizedIBProblem, ...)

Modified IBProblem to include base state. Only modification to the code is the direct product called by the nonlinear! function

Constructor

LinearizedIBProblem(
    base_state::IBState,
    base_prob::IBProblem,
    dt::Float64
)

Arugments

base_state::IBState:

base_prob::IBProblem:

dt::Float64: Time stepping interval.

IBPM.grid_of — Function

grid_of(x)

The Grid cooresponding to a problem or model.

IBPM.gridstep — Function

gridstep(x)

The minimum grid step size of a problem, model, or grid.

IBPM.model_of — Function

model_of(problem)

The SolnModel corresponding to problem.

IBPM.motions_of — Function

motions_of(x)

The Motion subtype corresponding to a problem or model.

IBPM.scheme_of — Function

scheme_of(problem)

The ExplicitScheme corresponding to problem.

IBPM.timestep — Function

timestep(x)

The time step size of a problem or scheme.