IBPM.AllTimesteps — Type

AllTimesteps() :: Timesteps

Specifies all the timesteps during a simulation.

IBPM.StateCallback — Type

StateCallback(f, timesteps::Timesteps, ret::Type=Any)

Specifies a callback f(state) that should be run at the given timesteps.

It is recommended to use each_timestep, at_times, or at_indices to construct this type. The return type of f can be specified with ret. The value of ret determines the type of StateResult that a callback is mapped to.

Can be passed to solve to specify callbacks and results.

IBPM.StateResult — Type

StateResult{T} <: AbstractVector{T}

The congregated results from a StateCallback. T is determined by the ret argument to StateCalllback.

The t field stores the time corresponding to each data entry. result[i] was taken at simulation time result.t[i].

IBPM.StopIteration — Type

StopIteration

Sentinel returned by a StateCallback that indicates that solve should terminate.

IBPM.TimestepIndices — Type

TimestepIndices(i) :: Timesteps

Specifies certain timestep indices during a simulation.

IBPM.TimestepTimes — Type

TimestepTimes(t) :: Timesteps

Specifies certain times during a simulation.

IBPM.Timesteps — Type

Timesteps

Specifies certain timesteps during a simulation.

IBPM.at_indices — Method

at_indices(f, i, [ret]) :: StateCallback

Return a StateCallback that runs f(state) at the n'th timestep for each n in i.

ret specifies the return type of f.

See also each_timestep, at_times.

IBPM.at_times — Method

at_times(f, t, [ret]) :: StateCallback

Return a StateCallback that runs f(state) at each time in t.

ret specifies the return type of f.

See also each_timestep, at_indices.

IBPM.compute_cfl — Method

compute_cfl(state, prob)

Compute the CFL number (uΔt/Δx) based on the fine-grid flux.

IBPM.each_timestep — Method

each_timestep(f, [ret]) :: StateCallback

Return a StateCallback that runs f(state) at each timestep.

ret specifies the return type of f.

See also at_times, at_indices.

IBPM.solve! — Method

solve!(state, problem, (t0, tf); [save], [call]) -> results

Solve problem between times t0 and tf by overwriting state each iteration.

See also solve.

IBPM.solve — Method

solve(problem, (t0, tf); [save], [call]) -> results

Solve problem between times t0 and tf, optionally saving or executing callbacks.

Arguments

problem::IBProblem: The problem specification.

(t0, tf): Initial and final times.

save: Collection of StateCallback that maps to results (default is empty).

call: Collection of StateCallback to be executed during the simulation.

Returns

The result of mapping each callback in save to a StateResult using map.