IBPM.advance! — Method

advance!(state::IBState, prob::AbstractIBProblem, t::Float64)

Advance state forward in time.

IBPM.advance! — Method

advance!(y::IBState, x::IBState, prob::AbstractIBProblem, t::Float64)

Advance state x forward in time and save in y

IBPM.advance — Method

advance(x::IBState, prob::AbstractIBProblem, t::Float64)

Advance state x forward in time (not mutating)

IBPM.base_flux! — Method

base_flux!(state::IBState, prob::AbstractIBProblem, t::Float64)

Set background flux based on prob.model.bodies[].motion

IBPM.base_flux! — Method

Update time-varying background flux for moving grid

IBPM.base_flux! — Method

Freestream flux for linearized problem (uniform zero)

IBPM.base_flux! — Method

Initialize irrotational freestream flux when not time-varying

IBPM.boundary_forces! — Method

boundary_forces!(F̃b, qs, q0, prob)

Solve the Poisson equation (25) in Colonius & Taira (2008).

Dispatch based on the type of motion in the problem - allows precomputing regularization and interpolation where possible.

IBPM.boundary_forces! — Method

boundary_forces!(::Type{V} where V <: Motion, F̃b, qs, q0, prob)

Solve the Poisson problem for bc2 = 0 (???) with nonzero boundary velocity ub

Bf̃ = Eq - ub
   = ECψ - ub

IBPM.boundary_forces! — Method

boundary_forces!(::Union{Type{Static}, Type{MovingGrid}},
                 F̃b, qs, q0, prob)

Solve modified Poisson problem for uB = 0 and bc2 = 0

 Bf̃ = Eq = ECψ

IBPM.get_trial_state! — Method

get_trial_state!(qs, Γs, state, prob)

Compute trial circulation Γs that doesn't satisfy no-slip BCs

Combine explicit Laplacian and nonlinear terms into a rhs then invert implicit part to return trial circulation Γs

High-level version of AB2: rhs = AΓ .- 3dt/2 * nonlin .+ dt/2 * nonlin_prev .+ dt/2 * rhsbc

Then do Ainv of that to back out trial circ

IBPM.project_circ! — Method

project_circ!(Γs, state, prob)

Update circulation to satisfy no-slip condition.

Dispatch based on the type of motion in the problem.

This allows precomputing regularization and interpolation where possible.

IBPM.update_bodies! — Method

update_bodies!(prob, t)

Update the immersed bodies and coupling matrices (if applicable).

IBPM.update_bodies! — Method

For moving grid, update the bodies and the coupling operators

IBPM.update_bodies! — Method

No motion for static or moving grid cases

IBPM.update_stress! — Method

update_stress!(state, prob)

Store surface stresses and integrated forces.

Mutates "state"